% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_table.R
\name{balance_table}
\alias{balance_table}
\title{Derive formatted balance table from matchit output}
\usage{
balance_table(matchit_out, threshold_smd = 0.2)
}
\arguments{
\item{matchit_out}{Output from matchit function}

\item{threshold_smd}{Threshold below which the absolute standardised mean difference is considered balanced (default = 0.2). If formal threshold is required, this should be set to NULL.}
}
\description{
Derive formatted balance table from matchit output
}
