% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalpsm.R
\name{finalpsm}
\alias{finalpsm}
\title{Wrapper for perfoming propensity-score matching with finalfit}
\usage{
finalpsm(matchit_out, dependent, explanatory = NULL, balance = T,
  metrics = T, fit = T)
}
\arguments{
\item{matchit_out}{Dataframe}

\item{dependent}{Character vector of length 1: quoted name of dependent variable. Can be continuous, a binary factor, or a survival object of form Surv(time, status).}

\item{explanatory}{Character vector of any length: quoted name(s) of explanatory variables (default = NULL - the strata and explanatory variables from the matchit_out object will be used)}

\item{balance}{Logical value whether a balance table should be supplied in the output (default = T).}

\item{metrics}{Logical value whether the model metrics should be supplied in the output (default = T).}

\item{fit}{Logical value whether the model fit object should be supplied in the output (default = T).}
}
\value{
Nested list of the final model table, and if specified the balance table, model metrics, and model fit.
}
\description{
matchit is the main command of the package MatchIt, which enables parametric models for causal inference to work better by selecting well-matched subsets of the original treated and control groups. This function acts as a wrapper to facilitate propensity-score matching with a tidyverse approach.
}
