% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Wrapper for the mice package}
\usage{
impute(
  data,
  impute,
  fixed = NULL,
  ignore = NULL,
  quiet = T,
  m = 5,
  maxit = 5,
  meth = "pmm",
  seed = 500,
  ...
)
}
\arguments{
\item{data}{Dataframe}

\item{impute}{Vector of column names containing variables for use in the imputation algorithm and for imputation.}

\item{fixed}{Optional vector of column names containing variables for use in the imputation algorithm but NOT for imputation.}

\item{ignore}{Optional vector of column names containing variables NOT for use in the imputation algorithm OR for imputation.}

\item{quiet}{Logical value to hide the output of the imputation process (default = T)}

\item{m}{The number of imputations to be performed (default = 5)}

\item{maxit}{is the number of iterations for each imputation  (see mice::mice() function).}

\item{seed}{random seed parameter which is useful for reproducibility.}

\item{...}{Additional arguments to be passed to the mice::mice() function.}

\item{method}{The imputation method to be used for each column in data (see mice::mice() function)}
}
\value{
Long dataset containing the original (unimputed) data, with each imputed dataset appended.
}
\description{
Impute missing at random data using the mice package in a tidyverse approach.
}
