% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputemutate.R
\name{imputemutate}
\alias{imputemutate}
\title{Wrapper for the mutate function to work across imputed datasets}
\usage{
imputemutate(.data, ...)
}
\arguments{
\item{...}{Mutations to perform on across all imputed datasets}

\item{data}{Output from impute function}
}
\value{
Same output from impute function with mutations applied
}
\description{
Modify imputed data across all datasets at once
}
