% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_table.R
\docType{data}
\name{matchit_out}
\alias{matchit_out}
\title{Derive formatted balance table from matchit output}
\format{An object of class \code{list} of length 2.}
\usage{
matchit_out
}
\arguments{
\item{matchit_out}{Output from matchit function}

\item{threshold_smd}{Threshold below which the absolute standardised mean difference is considered balanced (default = 0.2). If formal threshold is required, this should be set to NULL.}

\item{p}{Test for significant differences between treatment groups to assess balance before and after matching (default = FALSE)}
}
\description{
Derive formatted balance table from matchit output
}
\keyword{datasets}
