% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool2df.R
\name{pool2df}
\alias{pool2df}
\title{Pool imputed models together and output alongside unimputed model as fit2df tables.}
\usage{
pool2df(data, model = "model")
}
\arguments{
\item{data}{Dataframe containing 3 columns - imputed and m (from impute function) and a column containing nested model fits for each.}

\item{model}{Column name for the nested model fits within the tibble (default = "model")}
}
\value{
Returns a fit2df style table for (1) unimputed model (2) pooled imputed models (3) 1 and 2 joined.
}
\description{
Wrapper for performing finalfit modelling using imputed data
}
