% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_plot.R
\name{balance_plot}
\alias{balance_plot}
\title{Derive formatted balance table from matchit output}
\usage{
balance_plot(matchit_out, type = "covariate")
}
\arguments{
\item{matchit_out}{Output from matchit function}

\item{type}{Selection for plot being produced to allow visual assessment of balance. This includes a jitter or density plot (propensity score balance overall), and covariate plot (propensity score balance for individual variables)}
}
\description{
Derive formatted balance table from matchit output
}
