% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_table.R
\name{balance_table}
\alias{balance_table}
\title{Derive formatted balance table from matchit output}
\usage{
balance_table(matchit_out, threshold = 0.2, p = FALSE)
}
\arguments{
\item{matchit_out}{Output from matchit function}

\item{threshold}{Threshold below which the absolute standardised mean difference is considered balanced (default = 0.2). If formal threshold is required, this should be set to NULL.}

\item{p}{Test for significant differences between treatment groups to assess balance before and after matching (default = FALSE)}
}
\description{
Derive formatted balance table from matchit output
}
